+++ Assets/Scripts/Agent/StateMachine/IAgentStatus.cs
@@ -4,6 +4,6 @@ public interface IAgentStatus: IGameUpdate, IMeterHandler
 {
     void OnEnter(Dictionary<string,object> context);
     string GetStatusName();
-    void OnAction(int action);
+    void OnAction(byte action);
     void OnExit();
 }
